Finalizing style

+++ src/helpers.js
@@ -13,9 +13,9 @@ const toTimeInputValue = () => {
 /** @date ISO string dates */
 const ISODateStringToLocaleString = (date) => {
   let options = {
-    weekday: "long",
+    weekday: "short",
     year: "numeric",
-    month: "long",
+    month: "short",
     day: "numeric",
     hour: "2-digit",
     minute: "2-digit"
